requirements.txt
requests
tqdm
dateutils
regex
